% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartotools-package.R
\docType{package}
\name{cartotools-package}
\alias{cartotools-package}
\alias{cartotools}
\title{Simple tools I tend to use a lot when working with GIS data in R}
\description{
Feedback is very welcome!
}
\details{
\tabular{ll}{
	Package: \tab cartotools\cr
	Type: \tab Package\cr
	Version: \tab 0.0.0.9000\cr
	Date: \tab 2018-02-11\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach moritz.marbach@gess.ethz.ch
}
