% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr.R
\name{mbr}
\alias{mbr}
\title{Calculate the bounding box for a set of points}
\usage{
mbr(points)
}
\arguments{
\item{points}{a two-dimensional matrix of points}
}
\value{
\code{data.frame} of vertices of the minimum bounding rectangle
}
\description{
\code{mbr} uses a fast convex hull algorithm to calculate the 
minimum bounding rectangle.
}
\details{
The code was adapted from the code posted here
\url{https://gis.stackexchange.com/a/174577}
}
\examples{
\dontrun{

points <- matrix(rnorm(1000), ncol=2) 
box <- mbr(points)

plot(points,pch=20)
polygon(box,border='red')


}

}
\seealso{
\code{\link[spatstat]{boundingbox}}
}
