% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling.R
\name{tiling}
\alias{tiling}
\title{Slice raster image in square tiles}
\usage{
tiling(file, size, outsize = 100, path = NULL, co = "COMPRESS=JPEG",
  of = "GTiff", ...)
}
\arguments{
\item{file}{raster image file name}

\item{size}{size of square tiles in pixels}

\item{outsize}{reduction in resolution (default=none)}

\item{path}{path to write to}

\item{co}{compression algorithm}

\item{of}{output format}

\item{...}{further parameters passed to \code{gdal_translate}}
}
\value{
none
}
\description{
\code{tiling} uses \code{gdal_translate} to slice a raster
image into square tiles preserving the georeferencing of the input raster.
}
\details{
The default settings preserve the resolution but 
use JPEG compression to reduce the tiles' file size.
}
\examples{
\dontrun{

library(raster)

file <- system.file("extdata", "example.tif", package = "cartotools")
plot(raster(file))

tiling(file, size=10)
 
}

}
\seealso{
\code{\link[gdalUtils]{gdal_translate}}
}
